requirements.txt
blinker==1.8.2
cffi==1.16.0
click==8.1.7
cryptography==43.0.0
Flask==3.0.3
Flask-Cors==4.0.1
Flask-Simple-Crypt==0.3.3
itsdangerous==2.2.0
Jinja2==3.1.4
jsons==1.6.3
MarkupSafe==2.1.5
mysql-connector-python==9.0.0
pycparser==2.22
pycryptodome==3.20.0
PyMySQL==1.1.1
python-dotenv==1.0.1
SQLAlchemy==2.0.31
typing_extensions==4.12.2
typish==1.9.3
Werkzeug==3.0.3
